make okhttp client wrapper into a non implicit class
