+++ membership-common/src/main/scala/com/gu/lib/FutureHttpClient.scala
@@ -1,0 +1,7 @@
+package com.gu.lib
+
+import com.gu.monitoring.SafeLogging
+import okhttp3._
+
+import java.io.IOException
+import scala.concurrent.{Future, Promise}
